requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-telegram-bot==20.7
openai==1.12.0
python-dotenv
aiohttp
jinja2==3.1.3
python-jose[cryptography]
python-multipart==0.0.9
httpx~=0.25.2
pydantic>=2.0.0
starlette>=0.27.0
python-dateutil
supabase==2.3.1
passlib[bcrypt]>=1.7.4